+++ terraform/kubelet-private-registry/variables.tf
@@ -1,0 +1,9 @@
+variable "private_registries" {
+  description = "A list of private container-registries. Each entry must have an 'url', an 'username' and a 'password' (base64-encoded)"
+  type = list(object({
+    url = string
+    username = string
+    password = string
+  }))
+  sensitive   = true
+}