+++ CHANGELOG.md
@@ -39,3 +39,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Added missing KUBECONFIG export to setup
 - Gracefully shutdown k3s on vagrant halt/reload; #54
+- CES-Terraform-Module:
+  - Remove check if setup is applied because it needs a running cluster; #77